refactor: :recycle: refactor code. Take out components to new file
